-Correccion de errores del PullRequest
